+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow to provide files and boundary for multipart response.
 - Allow to provide data as a dictionary for multipart response.
 - Allow to provide callbacks that are executed upon reception of a request.
+- Handle the fact that parameters may be introduced in httpx *Dispatcher.send method.
 
 ### Changed
 - method can now be provided even if not entirely upper cased.